Extracted the "click" and "if" directives
