+++ src/js/modules/endev/endev.js
@@ -13,25 +13,6 @@ var endevModule = angular.module("Endev", ngModuleDeps);
 
 window.$injector = angular.injector(["ng","Endev"]);
 
-_.each([['if','ng-show'],['click','ng-click']],function(pair){
-  endevModule.directive(pair[0],['$compile',function($compile){
-    return {
-      terminal: true,
-      priority: 1000,
-      compile: function(element, attrs) {
-        element.attr(pair[1], attrs[pair[0]]);
-        element.removeAttr("data-" + pair[0]);
-        return {
-          pre: function preLink(scope, iElement, iAttrs, controller) {  },
-          post: function postLink(scope, iElement, iAttrs, controller) {  
-            $compile(iElement)(scope);
-          }
-        };
-      }
-    }
-  }]);
-},this);
-
 //The basic run
 endevModule.run(["$rootScope","$document","$templateCache",function($rootScope,$document,$templateCache){
   $rootScope.Date = Date;
@@ -66,11 +47,13 @@ require("./services/endevProvider")(endevModule);
 require("./services/endevRest")(endevModule);
 require("./services/endevYql")(endevModule);
 
+require("./directives/click")(endevModule);
 require("./directives/edit")(endevModule);
 require("./directives/else")(endevModule);
 require("./directives/endevAnnotation")(endevModule);
 require("./directives/endevItem")(endevModule);
 require("./directives/from")(endevModule);
+require("./directives/if")(endevModule);
 require("./directives/import")(endevModule);
 require("./directives/insertInto")(endevModule);
 require("./directives/list")(endevModule);
